components/qgis_plugin/pyproject.toml
[project]
name = "pinta-qgis-plugin"
requires-python = ">=3.12"
version = "0.0.0"
description = "QGIS plugin component of the elevation production system"
readme = "README.md"
dependencies = [
  "pinta-db",
  "qgis_plugin_tools",
]

[project.optional-dependencies]
# QGIS requirements are marked here to allow starting QGIS from venv without conflicting dependencies
qgis = [
  # "GDAL==3.8.4", # Cannot be installed via pip
  "numpy>=1, <2",
  # "matplotlib==3.6.3", # Slow to install, uncomment if we really need this...
  "OWSLib==0.29.3",
  "psycopg2-binary>=2, <3",
  "requests>=2, <3",
]
build = ["uv"]

[project.urls]
Homepage = "https://github.com/nlsfi/pinta"
Repository = "https://github.com/nlsfi/pinta"
Changelog = "https://github.com/nlsfi/pinta/blob/main/db/qgis_plugin/CHANGELOG.md/"

[build-system]
requires = ["uv_build>=0.10.9,<0.12.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "pinta_qgis_plugin"
include = [
  "py.typed",
  "metadata.txt",
  "resources/**/*",
  "**/*.svg",
  "**/*.ts",
  "**/*.ui",
  ".BUILD",
]

[tool.uv]
default-groups = ["dev", "lint", "qgis"]

[tool.uv.sources]
pinta-db = { workspace = true }

[dependency-groups]
dev = [
  "qgis-plugin-dev-tools>=0.11.0",
  # testing
  "pytest",
  "pytest-mock",
  "pytest-timeout",
  "pytest-dotenv",
  "coverage[toml]",
  "coverage-enable-subprocess",
  "pydevd-pycharm~=261.22158.340", # debugger for jetbrains ides
  "debugpy>=1.8.20",               # debugger for VSCode
]
lint = [
  # linting tools
  "prek",
  "mypy",
  "ruff",
  # flake8 plugins
  "flake8-qgis",
  # stubs for type checks
  "PyQt5-stubs",
  "types-requests",
]
# List all dev depencies that require Qt and/or QGIS here
qgis = [
  "pytest-qt",
  "pytest-qgis",
]

[tool.flake8]
select = [
  "QGS",
]


[tool.pytest.ini_options]
minversion = "7.0"
testpaths = [
  "test",
]
addopts = "--import-mode=importlib"

[tool.coverage.paths]
source = ["src", "**/site-packages"]

[tool.coverage.run]
source_pkgs = ["pinta_qgis_plugin"]
# generate distinct coverage files with relative paths
parallel = true
relative_files = false
# TODO: set branch=true and remove core="sysmon" when python 3.14
# is used (coverage will use sysmon by default)
branch = false
core = "sysmon"

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
]

# extend required for isort action to pick up correct config
[tool.ruff]
extend = "../../pyproject.toml"

[tool.qgis_plugin_dev_tools]
plugin_package_name = "pinta_qgis_plugin"
version_number_source = "distribution"
changelog_file_path = "../../CHANGELOG.md"
env_file_path = "../../.env"
runtime_requires = [
  "qgis_plugin_tools",
]
auto_add_recursive_runtime_dependencies = false
translation_language_codes = [
  "fi",
]
translation_search_paths = [
  "src/pinta_qgis_plugin",
]
translation_destination_path = "src/pinta_qgis_plugin/resources/i18n"


[tool.semantic_release]
commit_parser = "conventional-monorepo"
major_on_zero = false
allow_zero_version = true
commit_message = """\
chore(release): pinta-qgis-plugin@{version}

Automatically generated by python-semantic-release
"""
tag_format = "pinta-qgis-plugin-v{version}"
version_toml = ["pyproject.toml:project.version"]
build_command = """
    python -m pip install -e '.[build]'
    rm -f dist
    mkdir -p dist
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add ../../uv.lock
    uv build
    mv ../../dist/pinta_qgis_plugin* dist/
"""

[tool.semantic_release.commit_parser_options]
path_filters = ["."]
scope_prefix = "qgis_plugin"

[tool.semantic_release.changelog]
# Recommended patterns for conventional commits parser that is scope aware
exclude_commit_patterns = [
  '''chore(?:\([^)]*?\))?: .+''',
  '''ci(?:\([^)]*?\))?: .+''',
  '''refactor(?:\([^)]*?\))?: .+''',
  '''style(?:\([^)]*?\))?: .+''',
  '''test(?:\([^)]*?\))?: .+''',
  '''build\((?!deps\): .+)''',
  '''Initial [Cc]ommit.*''',
]

[tool.semantic_release.branches.main]
match = "(main|master|release-pipelines)" # TODO: remove this branch name
prerelease_token = "dev"
